Add JSON schema file for database's dances
The YAML that we use is simple enough that we can validate it with a
JSON schema. Most JSON schema validators know how to do that anyways.

The schema itself is written as YAML for readability. Again, it will be
fairly easy to find a validator that read this, or to convert it
beforehand.